requirements.txt
requests
python-dotenv
numpy
tf-nightly
h5py
discord.py
pillow